+++ frontend/src/types/createUserResponse.ts
@@ -1,6 +1,6 @@
 import type { User } from 'types/user'
 
-export type SignUpResponse = {
+export type CreateUserResponse = {
   user: User
   invitationToken: string
   isTeamCapacityReached: boolean
